+++ src/App.tsx
@@ -1,20 +1,18 @@
 import "./App.scss";
 import Header from "./components/Header/Header";
 import Hero from "./components/Hero/Hero";
-import Legend from "./components/Legend";
+import Legend from "./components/Legend/Legend";
 import Popular from "./components/PopularRestaurants/Popular";
 import SignatureDish from "./components/Signature-Dish/SignatureDish";
 
 function App() {
-  
-
   return (
     <div className="app">
       <Header />
       <Hero />
       <Popular />
-      < SignatureDish />
-      < Legend />
+      <SignatureDish />
+      <Legend />
     </div>
   );
 }
